requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
numpy
pandas
tiktoken
pydantic
PyPDF2
markdown
beautifulsoup4
openai>=1.40.0
streamlit
